FILE: moves RouterService to internal folder and removes its export from lib
